requirements.txt
flask
flask-apscheduler
flask-sqlalchemy
pytest
pytest-cov
pytest-mock
ruff
requests
requests-mock